requirements.txt
Django==1.7.2
Fabric==1.10.1
Pillow==2.7.0
Werkzeug==0.9.6
coverage==4.0a1
dj-database-url==0.3.0
dj-static==0.0.6
django-allauth==0.19.0
django-appconf==0.6
django-compressor==1.4
django-custom-user==0.5
django-debug-toolbar==1.2.2
django-extensions==1.4.9
django-filebrowser==3.5.7
django-grappelli==2.6.3
django-imagekit==3.2.5
django-nose==1.3
django-reversion==1.8.5
django-taggit==0.12.2
django-tinymce==1.5.3
django-toolbelt==0.0.1
ecdsa==0.11
flake8==2.3.0
gunicorn==19.1.1
heroku==0.1.4
ipython==2.3.1
mccabe==0.3
model-mommy==1.2.2
nose==1.3.4
nosexcover==1.0.10
oauthlib==0.7.2
paramiko==1.15.2
pep8==1.5.7
pilkit==1.1.12
psycopg2==2.5.4
pycrypto==2.6.1
pyflakes==0.8.1
pystache==0.5.4
python-dateutil==1.5
python-openid==2.2.5
raven==5.1.1
readline==6.2.4.1
requests==2.5.1
requests-oauthlib==0.4.2
six==1.7.3
sqlparse==0.1.14
static==1.1.1
static3==0.5.1
wsgiref==0.1.2
